pyproject.toml
[tool.poetry]
name = "anarchist-organization"
version = "0.1.0"
description = ""
authors = ["Samuel Panek <samuelrpanek@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.1"
django-rest-framework = "^0.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
